chore: split evm-stellar code into their own folders
