+++ src/services/blockchain/clients/mod.rs
@@ -3,8 +3,12 @@
 //! Contains specific implementations for different blockchain types:
 //! - EVM client for Ethereum-compatible chains
 //! - Stellar client for Stellar network
-mod evm;
-mod stellar;
+mod evm {
+	pub mod client;
+}
+mod stellar {
+	pub mod client;
+}
 
-pub use evm::{EvmClient, EvmClientTrait};
-pub use stellar::{StellarClient, StellarClientTrait};
+pub use evm::client::{EvmClient, EvmClientTrait};
+pub use stellar::client::{StellarClient, StellarClientTrait};
