+++ src/services/blockchain/transports/mod.rs
@@ -4,10 +4,13 @@
 //! - Web3 transport for EVM chains
 //! - Horizon and Stellar RPC transport for Stellar
 
-mod horizon;
-mod stellar;
-mod web3;
+mod evm {
+	pub mod web3;
+}
+mod stellar {
+	pub mod horizon;
+	pub mod soroban;
+}
 
-pub use horizon::HorizonTransportClient;
-pub use stellar::StellarTransportClient;
-pub use web3::Web3TransportClient;
+pub use evm::web3::Web3TransportClient;
+pub use stellar::{horizon::HorizonTransportClient, soroban::StellarTransportClient};
