+++ src/services/filter/filters/stellar/helpers.rs
