--- src/services/filter/helpers/mod.rs
@@ -1,7 +1,0 @@
-//! Helper functions for filter implementations.
-//!
-//! Contains chain-specific utility functions used by the filter
-//! implementations to process transactions and events.
-
-pub mod evm;
-pub mod stellar;
